backend/requirements.txt
fastapi
uvicorn
pydantic-ai
python-dotenv
openai
logfire
requests
